Fix problems detected by clang 19

+++ include/chesscore/move.h
@@ -31,7 +31,7 @@ struct Move {
     std::optional<Piece> promoted{};  ///< If the move promotes a piece, this is the promoted piece.
 
     CastlingRights castling_rights_before;            ///< Castling rights before the move.
-    size_t halfmove_clock_before;                     ///< Halfmove clock before the move.
+    int halfmove_clock_before;                        ///< Halfmove clock before the move.
     std::optional<Square> en_passant_target_before{}; ///< En passant target square before the move.
 
     /**
